// client/models/SubscriptionServiceDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SubscriptionServiceDto = {
    /**
     * Service ID
     */
    serviceId: string;
    /**
     * Number of times this service can be used
     */
    usageCount: number;
};